+++ R/Filter.R
@@ -24,7 +24,7 @@ filterData <- function(inputData, filterExpression, propagateDownwards = TRUE, p
 	sanitizeFilter <- function(filters) {
 		# Detect one or more "system" followed by 0 or one "2" and 0 or more spaces and then one or more parenthesis start:
 		usesSystem <- grepl("system+2? *\\(+", filters)
-		if(usesSystem) {
+		if(any(usesSystem)) {
 			stop("The following filter expression applies a call to the operating system, and may contain harmful code (please do not try to hack using StoX): ", filters)
 		}
 	}