+++ NEWS.md
@@ -2,6 +2,7 @@
 * Added sanitizeFilter() to avoid system calls in filter.
 * Remoevd hard coded conversions in ICESBiotic(), moving the responsibility of such conversions to the translation functions.
 * Optimized createOrderKey() for faster execution.
+* Renamed ReportICESAcoustic(), ReportICESBiotic() and ReportICESDatras() to WriteICESAcoustic(), WriteICESBiotic() and WriteICESDatras(), respectievly.
 
 # RstoxData v1.1.5 (2021-04-18)
 * Added TranslateICESAcoustic() and TranslateICESBiotic().